clean code for acm example data
